Be consistent with int|float vs float|int union types

+++ src/muqsit/arithmexp/operator/AdditionBinaryOperator.php
@@ -10,7 +10,7 @@ final class AdditionBinaryOperator extends SimpleBinaryOperator{
 		return new self("+", "Addition", BinaryOperatorPrecedence::ADDITION_SUBTRACTION, BinaryOperatorAssignmentType::LEFT);
 	}
 
-	public function operate(float|int $x, float|int $y) : int|float{
+	public function operate(int|float $x, int|float $y) : int|float{
 		return $x + $y;
 	}
 }
