+++ src/muqsit/arithmexp/operator/DivisionBinaryOperator.php
@@ -10,7 +10,7 @@ final class DivisionBinaryOperator extends SimpleBinaryOperator{
 		return new self("/", "Division", BinaryOperatorPrecedence::MULTIPLICATION_DIVISION_MODULO, BinaryOperatorAssignmentType::LEFT);
 	}
 
-	public function operate(float|int $x, float|int $y) : int|float{
+	public function operate(int|float $x, int|float $y) : int|float{
 		return $x / $y;
 	}
 }
