+++ src/muqsit/arithmexp/operator/MultiplicationBinaryOperator.php
@@ -10,7 +10,7 @@ final class MultiplicationBinaryOperator extends SimpleBinaryOperator{
 		return new self("*", "Multiplication", BinaryOperatorPrecedence::MULTIPLICATION_DIVISION_MODULO, BinaryOperatorAssignmentType::LEFT);
 	}
 
-	public function operate(float|int $x, float|int $y) : int|float{
+	public function operate(int|float $x, int|float $y) : int|float{
 		return $x * $y;
 	}
 }
